+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.0.11 (2021-05-27)
+
+ * Add code-inspector-check-quality that is similar
+   to code-inspector-github-action but for more Continuous Integration tools
+
+# 0.0.10 (2020-05-23)
+
+ * Add code-inspector-github-action
+
 # 0.0.1 (2019-01-10)
 
  * Initial version